+++ dcm_common/services/extensions/common.py
@@ -1,6 +1,8 @@
 """Common code for flask-extensions."""
 
 from typing import Iterable, Callable
+import sys
+import signal
 
 from flask import Flask
 
@@ -17,3 +19,21 @@ def startup_flask_run(
         _run(*args, **kwargs)
 
     app.run = _
+
+
+def add_signal_handler(signum: int, handler: Callable[[], None]) -> None:
+    """
+    Adds a `handler` for signal `signum` by prepending to the current
+    handler.
+    """
+
+    original_handler = signal.getsignal(signum)
+
+    def _handler(signum, frame):
+        """Stop threads, run original handler, and exit."""
+        handler()
+        if callable(original_handler):
+            original_handler(signum, frame)
+        sys.exit(0)
+
+    signal.signal(signum, _handler)